+++ ui/src/components/map/QasMap.vue
@@ -1,0 +1,93 @@
+<template>
+  <qas-box class="qas-map">
+    <!-- TODO descomentar quando implementar o input de pesquisa -->
+    <!-- <div v-if="hasSearch" class="items-center no-wrap row">
+      <gmap-autocomplete class="q-field__native q-placeholder" placeholder="Pesquisar..." />
+      <q-icon color="primary" name="o_search" size="24px" />
+    </div> -->
+    <gmap-map :center="centerPosition" class="qas-map__draw" :zoom="zoom">
+      <gmap-marker v-for="(marker, index) in markers" :key="index" :draggable="marker.draggable" :icon="marker.icon" :position="marker.position" @dragend="getPosition" @mouseout="closePopup" @mouseover="openPopup(marker, index)">
+        <gmap-info-window :opened="canShowPopup(index)">
+          <div class="text-weight-bold">{{ marker.title }}</div>
+          <div>{{ marker.description }}</div>
+        </gmap-info-window>
+      </gmap-marker>
+    </gmap-map>
+  </qas-box>
+</template>
+
+<script>
+import formMixin from '../../mixins/form.js'
+
+export default {
+  mixins: [
+    formMixin
+  ],
+
+  props: {
+    centerPosition: {
+      type: Object,
+      default: () => {}
+    },
+
+    hasSearch: {
+      type: Boolean
+    },
+
+    markers: {
+      type: Array,
+      default: () => []
+    },
+
+    showPopup: {
+      type: Boolean
+    },
+
+    zoom: {
+      type: Number,
+      default: 17
+    }
+  },
+
+  data () {
+    return {
+      isPopupDisplayed: false,
+      currentPlace: null,
+      indexMarker: ''
+    }
+  },
+
+  methods: {
+    openPopup ({ title, description }, index) {
+      this.indexMarker = index
+      this.isPopupDisplayed = !!(title || description)
+    },
+
+    canShowPopup (index) {
+      return this.isPopupDisplayed && this.showPopup && index === this.indexMarker
+    },
+
+    closePopup () {
+      this.isPopupDisplayed = false
+    },
+
+    getPosition (mouseEvent) {
+      this.$emit('update-position', mouseEvent.latLng.toJSON())
+    }
+
+    // TODO descomentar quando implementar o input de pesquisa
+    // setPlace (place) {
+    //   this.currentPlace = place
+    // }
+  }
+}
+</script>
+
+<style lang="scss">
+.qas-map {
+  &__draw{
+    height: 300px;
+    width: 100%;
+  }
+}
+</style>